+++ src/controllers/markdowneditorcontroller.cpp
@@ -70,6 +70,10 @@ void applyMarkdownConfigFields(const MarkdownEditorConfig &p_mdConfig,
 
   p_editorConfig->m_autoFoldPreviewedBlocksEnabled = p_mdConfig.getAutoFoldPreviewedBlocksEnabled();
 
+  if (!p_mdConfig.getConcealmentEnabled()) {
+    p_editorConfig->m_concealElements = vte::MarkdownConcealElement::None;
+  }
+
   // Map InplacePreviewSources flags from vnotex -> vte.
   {
     auto srcs = p_mdConfig.getInplacePreviewSources();