+++ src/widgets/dialogs/settings/markdowneditorpage.h
@@ -44,6 +44,8 @@ private:
 
   QCheckBox *m_autoSectionNumberInEditModeCheckBox = nullptr;
 
+  QCheckBox *m_concealmentCheckBox = nullptr;
+
   QCheckBox *m_constrainInplacePreviewWidthCheckBox = nullptr;
 
   QCheckBox *m_inplacePreviewSourceImageLinkCheckBox = nullptr;